LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/script-src-blocks-wasm.tentative.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://127.0.0.1/resources/execute-start.wasm because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://127.0.0.1/resources/execute-start.wasm because it does not appear in the script-src directive of the Content Security Policy.

PASS Importing a WebAssembly module should be guarded by script-src CSP.
